requirements.txt
numpy
torch
pygame
opencv2
tqdm